requirements.txt
google-cloud-vision
google-auth
python-dotenv
openai
playwright
imapclient
pyzmail36
